Organizando el modulo de HookTool
- Cambio del nombre del modulo a "HookTool"
- Se agregó el namespace a los script del "HookTool"
