--- Assets/Modules/Hook/HookAnchor.cs
@@ -1,13 +1,0 @@
-﻿using UnityEngine;
-
-[DisallowMultipleComponent]
-public class HookAnchor : MonoBehaviour
-{
-    public enum AnchorType
-    {
-        Swing,
-        Approach
-    }
-
-    public AnchorType typeOfAnchor;
-}
