+++ Assets/Modules/HookTool/HookAnchor.cs
@@ -1,0 +1,16 @@
+﻿using UnityEngine;
+
+namespace HookTool
+{
+    [DisallowMultipleComponent]
+    public class HookAnchor : MonoBehaviour
+    {
+        public enum AnchorType
+        {
+            Swing,
+            Approach
+        }
+
+        public AnchorType typeOfAnchor;
+    }
+}
