+++ Assets/Modules/HookTool/Icons/anchor_icon.png.meta
